pairing 4 is probably the solution
replace the existing code for creating new groups with this code. (Hopefully) this new code will first load all previous groups and convert the set of new groups to a tuple of tuples for comparison. Then, it will check if any previous groups have the same participants as new groups. If a duplicate group is found, it will reset new groups and try to create new groups again until successful. Finally, it will return the tuple of tuples of the new groups, which will then be used to output the groups as before.